pyproject.toml
[tool.poetry]
name = "template-code"
version = "0.1.0"
description = ""
authors = ["Thibaud ROBIN <trobin@esgi.com>"]
readme = "README.md"
packages = [
  { include = "src" },
  { include = "tp1", from = "src" },
  { include = "tp3", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
scapy = "^2.6.0"
pygal = "^3.0.5"
fpdf2 = "^2.8.1"
pytest = "^8.3.5"
capstone = "^5.0.6"
python-dotenv = "^1.2.1"
pylibemu = "^1.0"
openai = "^2.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tp1 = "tp1.main:main"
tp3 = "tp3.main:main"
